Sets up root and user reducers
