--- src/reducers/user.js
@@ -1,1 +1,0 @@
-// Esse reducer será responsável por tratar as informações da pessoa usuária
