--- src/reducers/wallet.js
@@ -1,1 +1,0 @@
-// Esse reducer será responsável por tratar o todas as informações relacionadas as despesas
